split facing into a separate file
